refactor: CellManagement 상태·로직을 useCellManagement 훅으로 분리
  - useCellManagement: 데이터 패칭, 아코디언, 체크, 다이얼로그 상태 관리
  - CellManagement 페이지는 UI 렌더링만 담당
