+++ src/pages/CellManagement.tsx
@@ -1,10 +1,7 @@
 import dayjs from "dayjs"
-import { useEffect, useState } from "react"
 import Layout from "../components/Layout/Layout"
 import useFadeIn from "../hooks/useFadeIn"
-import { McyMember } from "../types/McyMember"
-import { CellMember } from "../types/CellMember"
-import { getMcyMemberApi } from "../api/mcyMemberApi"
+import useCellManagement from "../hooks/useCellManagement"
 import CellManagementStats from "../components/cellManagement/CellManagementStats"
 import CellCard from "../components/cellManagement/CellCard"
 import MemberRow from "../components/cellManagement/MemberRow"
@@ -17,48 +14,22 @@ import MoveMembersDialog from "../components/cellManagement/MoveMembersDialog"
 const CellManagement = () => {
   const ref = useFadeIn()
   const year = dayjs().year()
-  const [cells, setCells] = useState<McyMember[]>([])
-  const [expandedCells, setExpandedCells] = useState<Set<string>>(new Set())
-  const [addCellOpen, setAddCellOpen] = useState(false)
-  const [addMemberOpen, setAddMemberOpen] = useState(false)
-  const [editTarget, setEditTarget] = useState<{ cellName: string; member: CellMember } | null>(null)
-  const [editCellTarget, setEditCellTarget] = useState<McyMember | null>(null)
-  const [moveOpen, setMoveOpen] = useState(false)
-  // Map<cellName, Set<memberName>>
-  const [checkedMembers, setCheckedMembers] = useState<Map<string, Set<string>>>(new Map())
-
-  const fetchCells = () => getMcyMemberApi().then(setCells)
-
-  useEffect(() => {
-    fetchCells()
-  }, [])
-
-  const totalChecked = Array.from(checkedMembers.values()).reduce((acc, s) => acc + s.size, 0)
-
-  const toggleCell = (cellName: string) => {
-    setExpandedCells(prev => {
-      const next = new Set(prev)
-      next.has(cellName) ? next.delete(cellName) : next.add(cellName)
-      return next
-    })
-  }
-
-  const toggleMemberCheck = (cellName: string, memberName: string) => {
-    setCheckedMembers(prev => {
-      const next = new Map(prev)
-      const cellSet = new Set(next.get(cellName) ?? [])
-      cellSet.has(memberName) ? cellSet.delete(memberName) : cellSet.add(memberName)
-      cellSet.size > 0 ? next.set(cellName, cellSet) : next.delete(cellName)
-      return next
-    })
-  }
-
-  const clearChecked = () => setCheckedMembers(new Map())
-
-  const handleMoveSuccess = () => {
-    clearChecked()
-    fetchCells()
-  }
+  const {
+    cells,
+    expandedCells,
+    addCellOpen, setAddCellOpen,
+    addMemberOpen, setAddMemberOpen,
+    editTarget, setEditTarget,
+    editCellTarget, setEditCellTarget,
+    moveOpen, setMoveOpen,
+    checkedMembers,
+    totalChecked,
+    fetchCells,
+    toggleCell,
+    toggleMemberCheck,
+    clearChecked,
+    handleMoveSuccess,
+  } = useCellManagement()
 
   return (
     <div className="w-full min-h-[100dvh] flex flex-col">
@@ -131,7 +102,6 @@ const CellManagement = () => {
         </section>
       </Layout>
 
-      {/* 멤버 선택 시 하단 액션 바 */}
       {totalChecked > 0 && (
         <div className="fixed bottom-0 left-0 right-0 px-5 pb-8 pt-3 bg-[#FFFCF6]/95 backdrop-blur-sm border-t border-black/5 flex items-center gap-3 z-40">
           <div className="flex-1">
